css/css-scroll-snap/snap-after-layout-change.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap/#snap-strictness" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
div {
  position: absolute;
  margin: 0px;
}

#scroller {
  height: 500px;
  width: 500px;
  overflow: hidden;
  scroll-snap-type: both mandatory;
}

#initial-target {
  width: 300px;
  height: 300px;
  top: 100px;
  left: 100px;
  background-color: blue;
  scroll-snap-align: start;
}

#other-target {
  width: 300px;
  height: 300px;
  top: 300px;
  left: 300px;
  background-color: green;
  scroll-snap-align: start;
}

.area {
  width: 2000px;
  height: 2000px;
}
</style>

<div id="scroller">
  <div class="area"></div>
  <div id="initial-target"></div>
  <div id="other-target"></div>
</div>

<script>
const initial_target = document.getElementById("initial-target");
const other_target = document.getElementById("other-target");
const scroller = document.getElementById("scroller");

test(() => {
  assert_equals(scroller.scrollTop, 100);
  assert_equals(scroller.scrollLeft, 100);
}, "Scroller should be snapped to the closest point on the first layout.");

test(() => {
  initial_target.style.setProperty("top", "400px");
  assert_equals(scroller.scrollTop, 400);
  assert_equals(scroller.scrollLeft, 100);
}, "Moving the current snap target should make the scroller resnap to it.");

test(() => {
  scroller.removeChild(initial_target);
  assert_equals(scroller.scrollTop, 300);
  assert_equals(scroller.scrollLeft, 300);
}, "Removing the current snap target should make the scroller snap to the other\
 target.");

test(() => {
  scroller.removeChild(other_target);
  scroller.appendChild(initial_target);
  assert_equals(scroller.scrollTop, 400);
  assert_equals(scroller.scrollLeft, 100);
}, "Adding a new snap area when there are none should make the scroller snap to\
 it.");

test(() => {
  scroller.appendChild(other_target);
  assert_equals(scroller.scrollTop, 400);
  assert_equals(scroller.scrollLeft, 100);
}, "Adding a new snap area while already snapped should not make the scroller \
snap.");
</script>
